fixed code, tried to make tests

+++ migrations/20190208085622368-base.js
@@ -95,7 +95,7 @@ var migrationCommands = [{
         'allowNull': false
       },
       'groups': {
-        'type': Sequelize.INTEGER(20),
+        'type': Sequelize.INTEGER,
         'field': 'groups'
       },
       'createdAt': {
